refactored earn similar to the Home route by using index.js files and moving css to css modules
