+++ src/components/modules/Earn/CurrentPoints/CurrentPoints.module.css
@@ -1,0 +1,5 @@
+.table {
+  margin: 5%;
+  width: 75%;
+  text-align: center;
+}
