+++ src/components/modules/Earn/CurrentPoints/index.js
@@ -1,0 +1,1 @@
+export { default } from "./CurrentPoints.js";
