+++ src/components/modules/Earn/EarnPoints/index.js
@@ -1,0 +1,1 @@
+export { default } from "./EarnPoints.js";
