+++ src/components/modules/Earn/SpendCard/index.js
@@ -1,0 +1,1 @@
+export { default } from "./SpendCard.js";
