+++ src/components/modules/Earn/SpendCardCollection/SpendCardCollection.module.css
@@ -1,0 +1,7 @@
+.collection {
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+  flex-wrap: wrap;
+}
