+++ src/components/modules/Earn/SpendCardCollection/index.js
@@ -1,0 +1,1 @@
+export { default } from "./SpendCardCollection.js";
