+++ src/components/modules/Earn/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Earn.js";
